samples: Bluetooth: encrypt llpm sample
It makes most sense to demonstrate llpm on an ecrypted link.
Also setting down conection interval to the smallest
bluetooth allows. 7.5ms. Since encyrpting the link takes
quite a lot of packets back and forth, this will speed up setup.
But also it makes sense that the sample compares the best
bluetooth can do with the best llpm can do.

Signed-off-by: Martin Tverdal <martin.tverdal@nordicsemi.no>