drivers: wifi: Don't set net config timeout to 0 implicitly
Using the default timeout value is beneficial in terms of support for
upstream sample, which often rely on network interface being up and
running before the sample code starts. Using the default timeout gives
the conn_mgr/supplicant a chance to connect to a Wi-Fi network before
the sample runs.

Therefore avoid modifying the default value of the timeout in the
driver. Instead, set the timeout explicitly to 0 for the NCS samples
that use the net config module.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>