samples: cellular: nrf_cloud_rest_device_message: use CAF
Updated the sample to use CAF instead of DK library for led
and button events. Added a device-tree overlay for LEDs to
work with CAF LEDs module. Added header files for buttons
and LEDs definition which are required by CAF.

Also, added error message for when the sample fails to send
alerts to nRF Cloud.

Signed-off-by: Tony Le <tony.le@nordicsemi.no>