lib: sms: Support for characters in hex format
Added support for giving data as binary but sending to network is
still always occuring in 7bit GSM encoding.
This is required by certification to send certain special characters
that are not available in ASCII and 0x00 which is '@' sign in
7bit GSM encoding.

Modem shell also updated to support new input data format.

"SMS subscriber" library renamed to "SMS" library.

Jira: MOSH-492

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>