pm: Fix shared SRAM handling
nRF5340 by default has 64KB shared SRAM enabled in the DTS independent
of whether IPC service to CPUNET is enabled or not. But when PM is
enabled it includes shared SRAM only when IPC service is enabled, this
wrongly causes the PM to include shared SRAM in to total available SRAM.

To fix, always include the PM configuration file for IPC shared memory
to account for the shared SRAM.

Fixes NCSDK-23152.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>