applications: asset_tracker_v2: Fix Wi-Fi+GNSS location priority
ATv2 is ignoring the result of the cloud location methods
(Wi-Fi and cellular) and simply indicating to Location library
that location is unknown. Hence, Location library will fallback to
next method.

This is fine when you have location method priority order as GNSS,
Wi-Fi and Cellular and there is nothing to fallback after WiFi and
cellular. However, when the priorities are Wi-Fi, GNSS and
Cellular, you don't want to fallback to GNSS if Wi-Fi positioning
returns a valid location.

Implemented support for waiting for the location from the cloud and
acting based on the outcome. If the cloud backend doesn't support
returning the location back to the device, Location library will be
still indicated that the location is unknown and it'll do fallback
to GNSS.

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>