samples: matter: Moved confirmation of the OTA image to the Init.
Confirmation of the OTA update should not be delegated to be
done in Matter Thread because Fprotect already has been set
during the factory data initialization and blocked writing
to the restricted areas. Confirming must be done in the
Init method before initializing the Factory data module.
Then, the result should be delivered to the image processor
implementation.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>