manifest: Update zephyr to fix broken dependency for Initial Attestation
Update zephyr to include changes that to fix the dependency on
Initial Attestation partition requiring an Initial Attestation Key in
order to boot.

Remove default enabling of Initial Attestation in non-Minimal
configuration. This default does not respect the original dependencies
and would enable the partition.

During merge the manifest update was dropped as other manifest update
moved the revision past this change.
Otherwise the manifest revision would have pointed to:
2ac74ca47b0ae06ff19755985e01b07b150f53b5

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>