tfm: Fix compilation error in provisioning with attestation enabled
Fix compilation error when TFM_NRF_PROVISIONING is enabled but the
INITIAL_ATTESTATION partition is disabled.
The TFM_NRF_PROVISIONING uses the SECURE_BOOT_STORAGE so select this
dependency.
The hal_attest.c file is used by both of these modules, so compile
the source file in when either is enabled.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>