applications: matter: Improve BLE reconnection mechanism for Bridge
Now the reconnection mechanism works independently of the normal
scan procedure.

The reconnection scan procedure is based on the lost BLE address,
not on UUID like for the normal scan.

When the connection to the BLE device is lost, the reconnection
mechanism starts working, but when a user wants to perform scanning
for the new devices, the mechanism stops working while the new scan
is in progress, and after that goes back to searching for the lost
device.

The scanned devices list is now independent of the reconnection
mechanism, so it is not lost during the reconnection scan in progress.

Also, connection to a new BLE device is possible while the
reconnection mechanism is working.

Signed-off-by: Arkadiusz Balys <arkadiusz.balys@nordicsemi.no>