samples: cellular: modem_shell: Improve "link edrx" command
Previously, the "link edrx" command required the system
(LTE-M/NB-IoT) to be selected, although eDRX was always enabled
for both systems.

The command syntax has now been changed so, that the system
parameters are no longer used. Instead, it's possible to give
eDRX and PTW values for both LTE-M and NB-IoT at the same time.

Signed-off-by: Tommi Kangas <tommi.kangas@nordicsemi.no>