wfa-qt: Disconnect STA before reset
Previously, we had removed the disconnect functionality, as we
considered `kernel reboot` to be similar to a disconnection.

However, in one of the SVD test cases, it appears that the
access point (AP) required a deauthentication frame to clear
specific buffers, enabling the proper sending and receiving of
ICMP traffic. As a result, ICMP traffic from the AP is observed
to start with sequence number 8 after a reset.

To address this issue, add the functionality that implicitly sends
a Deauthentication frame to the AP. This change aligns with the
behavior observed in the Linux case, where terminating the supplicant
also triggers the sending of a Deauthentication frame.

Fixes SHEL-2062

Signed-off-by: Triveni Danda <triveni.danda@nordicsemi.no>