applications: asset_tracker_v2: Cancelling previous location request
When there is no cloud connection, the data won’t be sent to cloud
but ATv2 is waiting for the cloud response. This causes location
library to wait for the feedback too and hence further location
requests are blocked until this ongoing request times out
(5mins by default).

This change checks if there is a cloud request pending when new
sampling starts and cancels previous location request.

Jira: NCSDK-24105

Signed-off-by: Tommi Rantanen <tommi.rantanen@nordicsemi.no>