sysbuild: adjust variant image to use existing variant kconfig impl
The ExternalProject_Add_Step() has an undesired side-effect.
The variant image would initially be using a regular generated .config
which would be replaced with the .config of the original image during
build.

Unfortunately the CMake configure of the variant image can in some cases
end with a CMake failure, which means the build stage is never reached
and thus the .config is never overwritten.

Fix this by copying the .config as part of the CMake configure stage
as is available in existing implementation.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>