nrf_security: Remove not useful PSA definitions
Remove PSA_CRYPTO_ACCELERATOR_DRIVER_PRESENT, this is always true as
all our drivers are considered "accelerators".

Remove PSA_CRYPTO_DRIVER_PRESENT, this appears to not have any valid
use.

Remove PSA_CRYPTO_SECURE, this is not used anyway and is not set
correctly currently, i.e. it is always false.
The configuration MBEDTLS_PSA_CRYPTO_SPM is the one that is being used.

NCSDK-22761

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>