samples: net: Remove settings backend override
The settings backend used to store the Azure hostname and device ID that
is received during DPS (Device Provisioning Service) was overridden by
the nRF7002 board configuration. Setting NVS as the backend instead
of FCB.

This caused the hostname and device ID to fail to be successfully
retrieved from flash post boot, causing the DPS process to occur
every time. Why it doesn't work with the NVS backend
needs to be investigated.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>