net: lib: wifi_mgmt_ext: Increase Wi-Fi connection retry to 30 seconds
Increase the fallback timeout for retrying a connection to stored SSIDs
to 20 seconds. The previous value of 10 seconds was just a bit too
little. And, would typically result in timeout in the WPA supplicant
after the device rebooted and attempted to connect to stored SSIDs.

Make the fallback timeout configurable.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>