nrf_security: Remove unused vendor max ECC bits
This code is not useful anymore for two reasons.
First of all the CONFIG_PSA_CRYPTO_DRIVER_*
symbols are now removed and so this will never
be executed.

The vendor max ECC bits are now configured
by the Oberon PSA core in the file:

core/include/psa/crypto_sizes.h

Depending on the PSA_WANT_* enabled curves.

Signed-off-by: Georgios Vasilakis <georgios.vasilakis@nordicsemi.no>