Bluetooth: Controller: Add VS command for setting max tx payload
Add the vendor specific command handler for
SDC_HCI_OPCODE_CMD_VS_MIN_VAL_OF_MAX_ACL_TX_PAYLOAD_SET.

Previously it was only possible to set a fixed value for this
using a Kconfig entry.

This command sets the minimum value of maximum ACL payload length
that can be sent in each packet. If the configured event length
is shorter than what is required to send a packet pair of 27 bytes
in each direction, the controller will use this value to determine
how much it can reduce the payload size to satisfy the event length
requirements LL Control PDUs are not affected by this API.

Signed-off-by: Rubin Gerritsen <rubin.gerritsen@nordicsemi.no>